refactor broadcast-table-raw-jdbc-example
