+++ sharding-jdbc-example/broadcast-table-example/broadcast-table-raw-jdbc-example/src/main/java/org/apache/shardingsphere/example/broadcast/table/raw/jdbc/JavaConfigurationExample.java
@@ -23,7 +23,8 @@ package org.apache.shardingsphere.example.broadcast.table.raw.jdbc;
 
 import org.apache.shardingsphere.example.broadcast.table.raw.jdbc.config.ShardingDatabasesConfiguration;
 import org.apache.shardingsphere.example.common.jdbc.repository.CountryRepositoryImpl;
-import org.apache.shardingsphere.example.common.jdbc.service.CountryServiceImpl;
+import org.apache.shardingsphere.example.common.jdbc.repository.SportsmanRepositoryImpl;
+import org.apache.shardingsphere.example.common.jdbc.service.SportsmanServiceImpl;
 import org.apache.shardingsphere.example.common.service.CommonService;
 
 import javax.sql.DataSource;
@@ -40,6 +41,6 @@ public class JavaConfigurationExample {
     }
 
     private static CommonService getCountryService(final DataSource dataSource) {
-        return new CountryServiceImpl(new CountryRepositoryImpl(dataSource));
+        return new SportsmanServiceImpl(new CountryRepositoryImpl(dataSource), new SportsmanRepositoryImpl(dataSource));
     }
 }
